NP-42898 moved ActionPanel content into own component
